requirements.txt
appdirs==1.4.4
asgiref==3.4.1
autopep8==1.5.7
black==21.7b0
click==8.0.1
Django==3.2.5
django-common-helpers==0.9.2
django-crontab==0.7.1
django-froala-editor==3.2.2
mypy-extensions==0.4.3
pathspec==0.9.0
pycodestyle==2.7.0
pytz==2021.1
regex==2021.7.6
sqlparse==0.4.1
toml==0.10.2
tomli==1.2.0
